+++ src/core/middleware/handlers/api/create-mock.handler.spec.ts
@@ -120,7 +120,7 @@ describe('CreateMocksHandler', () => {
         });
         it('shoud save the mock in the mocks folder', () => {
             handler.saveMock(mockPostData);
-            expect(outputJSONSync).toHaveBeenCalledWith(path.join('the/mocks/path', 'newname.somemock.json'), mockPostData, { spaces: 2 });
+            expect(outputJSONSync).toHaveBeenCalledWith(path.join(process.cwd(), '.ngapimock', 'generated', 'newname.mock.json'), mockPostData, { spaces: 2 });
         });
         it('should add a default response if no reponse is posted and save the mock', () => {
             mockPostData.responses = {};
@@ -135,7 +135,7 @@ describe('CreateMocksHandler', () => {
                     }
                 }
             };
-            expect(outputJSONSync).toHaveBeenCalledWith(path.join('the/mocks/path', 'newname.somemock.json'),
+            expect(outputJSONSync).toHaveBeenCalledWith(path.join(process.cwd(), '.ngapimock', 'generated', 'newname.mock.json'),
                 expectedPostData, { spaces: 2 });
         });
     });